Split SystemSettingsController into three smaller controllers
